Split propose settings stepper
